+++ config/utils.go
@@ -1,7 +1,8 @@
 package config
 
 import (
-	"fmt"
+	"encoding/json"
+	"log"
 	"strconv"
 	"strings"
 )
@@ -18,55 +19,110 @@ func contains(slice []string, sa string) bool {
 
 // getMostlyPrioriesConfigKey get key for the priories config value
 func getMostlyPrioriesConfigKey(option map[string]interface{}) interface{} {
+	// Parse arguments
 	if value, ok := option["arg"]; ok {
 		if option["type"] == "int" {
 			return *(value.(*int))
+
 		} else if option["type"] == "bool" {
 			return *(value.(*bool))
+
 		} else if option["type"] == "string_array" {
 			tmp := []string{}
 			for _, item := range *(value.(*StringArrayFlag)) {
 				tmp = append(tmp, item)
 			}
 			return tmp
+
+		} else if option["type"] == "rule_array" {
+			rules := []RuleConfig{}
+			for _, item := range *(value.(*StringArrayFlag)) {
+				rule := RuleConfig{}
+				err := json.Unmarshal([]byte(item), &rule)
+				if err != nil {
+					log.Panic(map[string]string{
+						"log_level": "panic",
+						"error":     "Rule unmarshal failed",
+						"for_rule":  item,
+						"details":   err.Error(),
+					})
+				}
+				rules = append(rules, rule)
+			}
+			return rules
+
+		} else if option["type"] == "string" {
+			return *(value.(*string))
 		}
-		return *(value.(*string))
+
+		// Parse config file
 	} else if value, ok := option["file"]; ok {
 		if option["type"] == "int" {
 			i, _ := strconv.Atoi(value.(string))
 			return i
+
 		} else if option["type"] == "bool" {
 			return value.(string) == "true"
+
 		} else if option["type"] == "string_array" {
-			valueSlice := value.([]interface{})
-			stringSlice := []string{}
-			for _, v := range valueSlice {
-				stringSlice = append(stringSlice, v.(string))
+			return InterfaceToStringSlice(value)
+
+		} else if option["type"] == "rule_array" {
+			rules := []RuleConfig{}
+			for _, v := range value.([]interface{}) {
+				rule := RuleConfig{}
+				rule.FromMap(v.(map[interface{}]interface{}))
+				rules = append(rules, rule)
 			}
-			return stringSlice
+			return rules
+
+		} else if option["type"] == "string" {
+			return value.([]map[string]string)
 		}
-		return value
+
+		// Parse environment
 	} else if value, ok := option["env"]; ok {
 		if option["type"] == "int" {
 			i, _ := strconv.Atoi(value.(string))
 			return i
+
 		} else if option["type"] == "bool" {
 			return value.(string) == "true"
+
 		} else if option["type"] == "string_array" {
 			return strings.Split(value.(string), ",")
+
+		} else if option["type"] == "rule_array" {
+			rules := []RuleConfig{}
+			err := json.Unmarshal([]byte(value.(string)), &rules)
+			if err != nil {
+				log.Panic(map[string]string{
+					"log_level": "panic",
+					"error":     "Rules unmarshal failed",
+					"for_rule":  value.(string),
+					"details":   err.Error(),
+				})
+			}
+			return rules
+
+		} else if option["type"] == "string" {
+			return value.(string)
 		}
-		return value
 	}
 
 	// Send back default
 	if option["type"] == "string_array" {
-		fmt.Println(option["default"].(*StringArrayFlag))
 		tmp := []string{}
 		for _, item := range *(option["default"].(*StringArrayFlag)) {
 			tmp = append(tmp, item)
 		}
 		return tmp
 	}
+
+	if option["type"] == "rule_array" {
+		return []RuleConfig{}
+	}
+
 	return option["default"]
 }
 
@@ -82,4 +138,13 @@ func (i *StringArrayFlag) String() string {
 func (i *StringArrayFlag) Set(value string) error {
 	*i = append(*i, value)
 	return nil
+}
+
+// InterfaceToStringSlice convert a interface to string slice
+func InterfaceToStringSlice(i interface{}) []string {
+	stringSlice := []string{}
+	for _, v := range i.([]interface{}) {
+		stringSlice = append(stringSlice, v.(string))
+	}
+	return stringSlice
 }