+++ rules/whitelist.go
@@ -5,26 +5,24 @@ import (
 	"auth-guardian/logging"
 	"fmt"
 	"net/http"
-	"regexp"
 )
 
-var whitelistRules []*regexp.Regexp
+var whitelistRules []Rule
 
 // InitializeWhitelistMiddleware initialize the whitelist middleware
 func InitializeWhitelistMiddleware() {
 	logging.Debug(&map[string]string{"file": "whitelist.go", "Function": "InitializeWhitelistMiddleware", "event": "Initialize whitelist middleware"})
 
 	for _, rule := range config.WhitelistRules {
-		reg, err := regexp.Compile(rule)
-		if err != nil {
-			logging.Fatal(&map[string]string{
-				"file":         "whitelist.go",
-				"Function":     "InitializeWhitelistMiddleware",
-				"event":        "Whitelist regex parsing failed",
-				"regex_string": rule,
-			})
-		}
-		whitelistRules = append(whitelistRules, reg)
+		ruleStruct := Rule{}
+		ruleStruct.FromRuleConfig(rule)
+		whitelistRules = append(whitelistRules, ruleStruct)
+
+		logging.Info(&map[string]string{
+			"event":       "Rule added",
+			"rule_method": fmt.Sprintf("%v", rule.Method),
+			"rule_path":   rule.Path,
+		})
 	}
 }
 
@@ -32,28 +30,33 @@ func InitializeWhitelistMiddleware() {
 func WhitelistRuleMiddleware(nextSecured http.Handler, nextWhitelisted http.Handler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		logging.Debug(&map[string]string{
-			"file":       "whitelist.go",
-			"Function":   "WhitelistRuleMiddleware",
-			"event":      "Handle request",
-			"method":     r.Method,
-			"req_scheme": r.URL.Scheme,
-			"req_host":   r.Host,
-			"req_path":   r.URL.Path,
-			"req_query":  r.URL.RawQuery,
+			"file":      "whitelist.go",
+			"Function":  "WhitelistRuleMiddleware",
+			"event":     "Handle request",
+			"method":    r.Method,
+			"req_path":  r.URL.Path,
+			"req_query": r.URL.RawQuery,
 		})
-
-		// Build request string
-		requestString := fmt.Sprintf("%s %s", r.Method, r.URL.Path)
 
 		// Check if request match whitelist rule
 		for _, rule := range whitelistRules {
-			if rule.MatchString(requestString) {
+			if rule.TestWhitelist(r) {
+				logging.Debug(&map[string]string{
+					"file":     "whitelist.go",
+					"Function": "WhitelistRuleMiddleware",
+					"event":    "Serve to upstream",
+				})
 				nextWhitelisted.ServeHTTP(w, r)
 				return
 			}
 		}
 
 		// Serve to auth middleware
+		logging.Debug(&map[string]string{
+			"file":     "whitelist.go",
+			"Function": "WhitelistRuleMiddleware",
+			"event":    "Serve to auth middleware",
+		})
 		nextSecured.ServeHTTP(w, r)
 	})
 }